requirements.txt
apache-airflow
psycopg2-binary
boto3
pandas
pyarrow
